LayoutTests/accessibility/range-alter-by-step.html

<!DOCTYPE HTML PUBLIC "-//IETF//DTD HTML//EN">
<html>
<head>
<script src="../resources/js-test.js"></script>
<script src="../resources/accessibility-helper.js"></script>
</head>
<body>

<input type="range" min="0" max="100" value="25" step="25" id="rangeInput"/>

<script>
    description("This test makes sure that if a range type has a step value, that increment and decrement work.");

    var rangeInput;
    async function waitForAndExpectIntValue(expectedIntValue) {
        await waitForExpression(rangeInput, "intValue", expectedIntValue);
        shouldBe("rangeInput.intValue", `${expectedIntValue}`)
    }

    if (window.accessibilityController) {
        window.jsTestIsAsync = true;

        setTimeout(async function() {
            // ARIA determinate progressbar
            rangeInput = await waitForElementById("rangeInput");
            await waitForAndExpectIntValue(25);

            rangeInput.increment();
            await waitForAndExpectIntValue(50);

            rangeInput.increment();
            await waitForAndExpectIntValue(75);

            rangeInput.increment();
            await waitForAndExpectIntValue(100);

            rangeInput.increment();
            await waitForAndExpectIntValue(100);

            rangeInput.decrement();
            await waitForAndExpectIntValue(75);

            rangeInput.decrement();
            await waitForAndExpectIntValue(50);

            rangeInput.decrement();
            await waitForAndExpectIntValue(25);

            rangeInput.decrement();
            await waitForAndExpectIntValue(0);

            rangeInput.decrement();
            await waitForAndExpectIntValue(0);

            finishJSTest();
        }, 0);
    }
</script>
</body>
</html>
